kafka-health-agent/pyproject.toml
[project]
name = "kafka-health-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.26.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
